pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "black[jupyter]>=25.1.0",
    "cartopy>=0.24.1",
    "dask>=2025.1.0",
    "geopandas>=1.0.1",
    "gitpython>=3.1.44",
    "ipykernel>=6.29.5",
    "llvmlite>=0.44.0",
    "matplotlib>=3.10.0",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pip>=25.0",
    "pre-commit>=4.1.0",
    "pyarrow>=19.0.0",
    "rasterio>=1.4.3",
    "rioxarray>=0.18.2",
    "scipy>=1.15.1",
    "uv>=0.5.26",
    "xarray-spatial>=0.4.0",
]
